refactor RegImage
